Change the base class of enemy to character

+++ btp600Assignment2/src/btp600Assignment2/Enemies/ObserverMinion.java
@@ -1,5 +1,5 @@
 package btp600Assignment2.Enemies;
 
 public interface ObserverMinion {
-	public void update(int hp, int minionAP);
+	public void update(long hp, long minionAP);
 }
